# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/document/create": {
      "post": {
        "operationId": "DocumentController_createDocument",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDocumentDto"
                }
              }
            }
          }
        },
        "tags": [
          "document"
        ]
      }
    },
...
    "/document/{id}": {
      "get": {
        "operationId": "DocumentController_getDocumentById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Document ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Document by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDocumentDto"
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          }
        },
        "tags": [
          "document"
        ]
      }
    }
  },
  "info": {
    "title": "Stepture API",
    "description": "API documentation for Stepture Backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateDocumentDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the document",
            "example": "How to Login to Admin Dashboard",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "description": "Optional detailed description of what this document covers",
            "example": "A comprehensive step-by-step guide for new administrators to access the dashboard",
            "nullable": true,
            "maxLength": 1000
          },
          "steps": {
            "description": "Array of steps that make up this document",
            "minItems": 1,
            "maxItems": 100,
            "example": [
              {
                "stepNumber": 1,
                "stepDescription": "Navigate to the login page at https://admin.example.com",
                "type": "STEP",
                "screenshot": {
                  "url": "https://example.com/screenshots/login-page.png",
                  "viewportX": 1920,
                  "viewportY": 1080,
                  "pageX": 0,
                  "pageY": 0,
                  "devicePixelRatio": 1
                }
              },
              {
                "stepNumber": 2,
                "stepDescription": "Remember to use your corporate email address",
                "type": "TIPS"
              },
              {
                "stepNumber": 3,
                "stepDescription": "Enter your email address in the username field",
                "type": "STEP",
                "screenshot": {
                  "url": "https://example.com/screenshots/username-field.png",
                  "viewportX": 1920,
                  "viewportY": 1080,
                  "pageX": 0,
                  "pageY": 100,
                  "devicePixelRatio": 1
                }
              },
              {
                "stepNumber": 4,
                "stepDescription": "Enter your password in the password field",
                "type": "STEP"
              },
              {
                "stepNumber": 5,
                "stepDescription": "Make sure your password meets the security requirements",
                "type": "ALERT"
              },
              {
                "stepNumber": 6,
                "stepDescription": "Click the \"Sign In\" button to authenticate",
                "type": "STEP",
                "screenshot": {
                  "url": "https://example.com/screenshots/signin-button.png",
                  "viewportX": 1920,
                  "viewportY": 1080,
                  "pageX": 0,
                  "pageY": 200,
                  "devicePixelRatio": 1
                }
              }
            ],
            "items": {
              "type": "array"
            },
            "type": "array"
          }
        },
        "required": [
          "title",
          "steps"
        ]
      }
    }
  }
}